+++ tests/unit/conftest.py
@@ -3,14 +3,15 @@ File to test the encodings for the Leela Chess Zero engine.
 """
 
 import random
-
 import chess
 import pytest
+
+from lczerolens import LczeroBoard
 
 
 @pytest.fixture(scope="module")
 def random_move_board_list():
-    board = chess.Board()
+    board = LczeroBoard()
     seed = 42
     random.seed(seed)
     move_list = []
@@ -25,7 +26,7 @@ def random_move_board_list():
 
 @pytest.fixture(scope="module")
 def repetition_move_board_list():
-    board = chess.Board()
+    board = LczeroBoard()
     move_list = []
     board_list = [board.copy()]
     for uci_move in ("b1a3", "b8c6", "a3b1", "c6b8") * 4:
@@ -38,7 +39,7 @@ def repetition_move_board_list():
 
 @pytest.fixture(scope="module")
 def long_move_board_list():
-    board = chess.Board()
+    board = LczeroBoard()
     seed = 6
     random.seed(seed)
     move_list = []
