+++ tests/unit/lenses/test_activation.py
