+++ tests/unit/lenses/test_policy.py
