+++ tests/unit/play/test_sampling.py
@@ -1,14 +1,13 @@
 """Sampling tests."""
 
-import chess
-
 from lczerolens.play.sampling import ModelSampler, SelfPlay, PolicySampler, RandomSampler
+from lczerolens.board import LczeroBoard
 
 
 class TestRandomSampler:
     def test_get_utilities(self):
         """Test get_utilities method."""
-        board = chess.Board()
+        board = LczeroBoard()
         sampler = RandomSampler()
         utility, _, _ = next(iter(sampler.get_utilities([board, board])))
         assert utility.shape[0] == 20
@@ -17,21 +16,21 @@ class TestRandomSampler:
 class TestModelSampler:
     def test_get_utilities_tiny(self, tiny_model):
         """Test get_utilities method."""
-        board = chess.Board()
+        board = LczeroBoard()
         sampler = ModelSampler(tiny_model, use_argmax=False)
         utility, _, _ = next(iter(sampler.get_utilities([board, board])))
         assert utility.shape[0] == 20
 
     def test_get_utilities_winner(self, winner_model):
         """Test get_utilities method."""
-        board = chess.Board()
+        board = LczeroBoard()
         sampler = ModelSampler(winner_model, use_argmax=False)
         utility, _, _ = next(iter(sampler.get_utilities([board, board])))
         assert utility.shape[0] == 20
 
     def test_policy_sampler_tiny(self, tiny_model):
         """Test policy_sampler method."""
-        board = chess.Board()
+        board = LczeroBoard()
         sampler = PolicySampler(tiny_model, use_argmax=False)
         utility, _, _ = next(iter(sampler.get_utilities([board, board])))
         assert utility.shape[0] == 20
@@ -40,7 +39,7 @@ class TestModelSampler:
 class TestSelfPlay:
     def test_play(self, tiny_model, winner_model):
         """Test play method."""
-        board = chess.Board()
+        board = LczeroBoard()
         white = ModelSampler(tiny_model, use_argmax=False)
         black = ModelSampler(winner_model, use_argmax=False)
         self_play = SelfPlay(white=white, black=black)
@@ -57,7 +56,7 @@ class TestSelfPlay:
 class TestBatchedPolicySampler:
     def test_batched_policy_sampler_ag(self, tiny_model):
         """Test batched_policy_sampler method."""
-        boards = [chess.Board() for _ in range(10)]
+        boards = [LczeroBoard() for _ in range(10)]
 
         sampler_ag = PolicySampler(tiny_model, use_argmax=True)
         moves = sampler_ag.get_next_moves(boards)
@@ -66,7 +65,7 @@ class TestBatchedPolicySampler:
 
     def test_batched_policy_sampler_no_ag(self, tiny_model):
         """Test batched_policy_sampler method."""
-        boards = [chess.Board() for _ in range(10)]
+        boards = [LczeroBoard() for _ in range(10)]
 
         sampler_no_ag = PolicySampler(tiny_model, use_argmax=False)
         moves = sampler_no_ag.get_next_moves(boards)
@@ -74,7 +73,7 @@ class TestBatchedPolicySampler:
 
     def test_batched_policy_sampler_no_ag_sub(self, tiny_model):
         """Test batched_policy_sampler method."""
-        boards = [chess.Board() for _ in range(10)]
+        boards = [LczeroBoard() for _ in range(10)]
 
         sampler_no_ag = PolicySampler(tiny_model, use_argmax=False, use_suboptimal=True)
         moves = sampler_no_ag.get_next_moves(boards)
